requirements.txt
fastapi
uvicorn
pydantic
motor
pandas
python-dotenv
pymongo
python-multipart
jinja2